+++ checkpoint/CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- Asynchronous directory creation enabled by default.
+
 ## [0.11.11] - 2025-04-07
 
 ### Added